Refactor AlarmService to use AlarmPersistenceService
